requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
motor==3.1.2
pydantic==1.10.7
python-dotenv==1.0.0
